+++ src/GRA.Domain/Entities/Orcamento.cs
@@ -4,10 +4,10 @@ namespace GRA.Domain.Entities;
 
 public class Orcamento : Entity
 {
-    public int OficinaId { get; set; }
+    public long OficinaId { get; set; }
     public Oficina? Oficina { get; set; }
 
-    public int OrdemServicoId { get; set; }
+    public long OrdemServicoId { get; set; }
     public OrdemServico? OrdemServico { get; set; }
 
     public DateTime DataCriacao { get; set; } = DateTime.UtcNow;
